+++ SimPEG/EM/FDEM/__init__.py
@@ -1,5 +1,5 @@
 from SurveyFDEM import Survey
 import SrcFDEM as Src
 import RxFDEM as Rx
-from FDEM import Problem3D_e, Problem3D_b, Problem3D_j, Problem3D_h
+from ProblemFDEM import Problem3D_e, Problem3D_b, Problem3D_j, Problem3D_h
 from FieldsFDEM import Fields3D_e, Fields3D_b, Fields3D_j, Fields3D_h
